+++ composeApp/src/commonMain/kotlin/it/homebudget/app/ui/screens/dashboard/DashboardScaffold.kt
@@ -19,7 +19,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.ionspin.kotlin.bignum.integer.BigInteger
 import it.homebudget.app.ui.screens.AndroidDataTransferUi
 import it.homebudget.app.ui.screens.AndroidNavigationDestination
 import it.homebudget.app.ui.screens.AndroidNavigationRailOverlay
@@ -34,7 +33,7 @@ internal fun DashboardScreenScaffold(
     strings: DashboardStrings,
     openVoiceExpenseRequest: Int = 0,
     selectedMonth: MonthCursor,
-    totalAmount: BigInteger,
+    totalAmount: Long,
     showFab: Boolean,
     onOpenCategories: () -> Unit,
     onOpenAddExpense: () -> Unit,
